requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-otp==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
